tests

+++ SpinnerButton/Classes/SpinnerButton.h
@@ -21,11 +21,11 @@ IB_DESIGNABLE
 /// Stroke line width of the animated border (default: 5)
 @property (nonatomic, assign) IBInspectable CGFloat strokeLineWidth;
 
-/// Corner radius (default: 10)
+/// Define a corner radius (default: 10)
 @property (nonatomic, assign) IBInspectable CGFloat cornerRadius;
 
 /// Time interval of animation duration
-@property (nonatomic, assign) IBInspectable int timeInterval;
+@property (nonatomic, assign) IBInspectable double timeInterval;
 
 #pragma mark - Methods
